+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/xplat/RegistryWrapper.kt
@@ -10,4 +10,6 @@ interface RegistryWrapper<T> : HolderGetter<T>, Iterable<T> {
 
     fun tryGet(location: ResourceLocation): T?
     fun get(id: Int): T
+
+    fun keySet(): Set<ResourceLocation>
 }