+++ projects/forge/src/main/kotlin/site/siredvin/peripheralium/forge/ForgePeripheraliumPlatform.kt
@@ -88,6 +88,10 @@ object ForgePeripheraliumPlatform : PeripheraliumPlatform {
         override fun iterator(): Iterator<T> {
             return registry.iterator()
         }
+
+        override fun keySet(): Set<ResourceLocation> {
+            return registry.keys
+        }
     }
 
     override val fluidCompactDivider: Int